refactor: `_speaker_info` を `MetasStore` へ移植
