+++ voicevox_engine/metas/MetasStore.py
@@ -1,21 +1,21 @@
+import base64
 import json
 from copy import deepcopy
 from pathlib import Path
-from typing import TYPE_CHECKING, Literal, NewType
+from typing import Literal, NewType
 
-from pydantic import BaseModel, Field
+from fastapi import HTTPException
+from pydantic import BaseModel, Field, parse_obj_as
 
+from voicevox_engine.core.core_adapter import CoreAdapter
 from voicevox_engine.metas.Metas import (
     Speaker,
+    SpeakerInfo,
     SpeakerStyle,
     SpeakerSupportedFeatures,
     StyleId,
     StyleType,
 )
-
-if TYPE_CHECKING:
-    from voicevox_engine.core.core_adapter import CoreAdapter
-
 
 _CoreStyleId = NewType("_CoreStyleId", int)
 _CoreStyleType = Literal["talk", "singing_teacher", "frame_decode", "sing"]
@@ -60,6 +60,10 @@ class _EngineSpeaker(BaseModel):
     )
 
 
+def b64encode_str(s: bytes) -> str:
+    return base64.b64encode(s).decode("utf-8")
+
+
 class MetasStore:
     """
     話者やスタイルのメタ情報を管理する
@@ -72,6 +76,7 @@ class MetasStore:
         engine_speakers_path : Path
             エンジンに含まれる話者メタ情報ディレクトリのパス。
         """
+        self._speakers_path = engine_speakers_path
         # エンジンに含まれる各話者のメタ情報
         self._loaded_metas: dict[str, _EngineSpeaker] = {
             folder.name: _EngineSpeaker(
@@ -109,6 +114,101 @@ class MetasStore:
             )
             for speaker_meta in core_metas
         ]
+
+    def speaker_info(
+        self,
+        speaker_uuid: str,
+        speaker_or_singer: Literal["speaker", "singer"],
+        core: CoreAdapter,
+    ) -> SpeakerInfo:
+        # エンジンに含まれる話者メタ情報は、次のディレクトリ構造に従わなければならない：
+        # {root_dir}/
+        #   speaker_info/
+        #       {speaker_uuid_0}/
+        #           policy.md
+        #           portrait.png
+        #           icons/
+        #               {id_0}.png
+        #               {id_1}.png
+        #               ...
+        #           portraits/
+        #               {id_0}.png
+        #               {id_1}.png
+        #               ...
+        #           voice_samples/
+        #               {id_0}_001.wav
+        #               {id_0}_002.wav
+        #               {id_0}_003.wav
+        #               {id_1}_001.wav
+        #               ...
+        #       {speaker_uuid_1}/
+        #           ...
+
+        # 該当話者の検索
+        speakers = parse_obj_as(list[Speaker], json.loads(core.speakers))
+        speakers = filter_speakers_and_styles(speakers, speaker_or_singer)
+        for i in range(len(speakers)):
+            if speakers[i].speaker_uuid == speaker_uuid:
+                speaker = speakers[i]
+                break
+        else:
+            # FIXME: ドメインを合わせる
+            raise HTTPException(status_code=404, detail="該当する話者が見つかりません")
+
+        try:
+            speaker_path = self._speakers_path / speaker_uuid
+            # 話者情報の取得
+            # speaker policy
+            policy_path = speaker_path / "policy.md"
+            policy = policy_path.read_text("utf-8")
+            # speaker portrait
+            portrait_path = speaker_path / "portrait.png"
+            portrait = b64encode_str(portrait_path.read_bytes())
+            # スタイル情報の取得
+            style_infos = []
+            for style in speaker.styles:
+                id = style.id
+                # style icon
+                style_icon_path = speaker_path / "icons" / f"{id}.png"
+                icon = b64encode_str(style_icon_path.read_bytes())
+                # style portrait
+                style_portrait_path = speaker_path / "portraits" / f"{id}.png"
+                style_portrait = None
+                if style_portrait_path.exists():
+                    style_portrait = b64encode_str(style_portrait_path.read_bytes())
+                # voice samples
+                voice_samples = [
+                    b64encode_str(
+                        (
+                            speaker_path
+                            / "voice_samples/{}_{}.wav".format(id, str(j + 1).zfill(3))
+                        ).read_bytes()
+                    )
+                    for j in range(3)
+                ]
+                style_infos.append(
+                    {
+                        "id": id,
+                        "icon": icon,
+                        "portrait": style_portrait,
+                        "voice_samples": voice_samples,
+                    }
+                )
+        except FileNotFoundError:
+            import traceback
+
+            traceback.print_exc()
+            # FIXME: ドメインを合わせる
+            raise HTTPException(
+                status_code=500, detail="追加情報が見つかりませんでした"
+            )
+
+        ret_data = SpeakerInfo(
+            policy=policy,
+            portrait=portrait,
+            style_infos=style_infos,
+        )
+        return ret_data
 
 
 def construct_lookup(
